+++ table/src/main/java/tech/ydb/table/settings/PartitioningSettings.java
@@ -9,6 +9,10 @@ public class PartitioningSettings {
     private Boolean partitioningByLoad;
     @Nullable
     private Long partitionSizeMb;
+    @Nullable
+    private Long minPartitionsCount;
+    @Nullable
+    private Long maxPartitionsCount;
 
     @Nullable
     public Boolean getPartitioningBySize() {
@@ -25,6 +29,16 @@ public class PartitioningSettings {
         return partitionSizeMb;
     }
 
+    @Nullable
+    public Long getMinPartitionsCount() {
+        return minPartitionsCount;
+    }
+
+    @Nullable
+    public Long getMaxPartitionsCount() {
+        return maxPartitionsCount;
+    }
+
     public PartitioningSettings setPartitioningBySize(boolean partitioningBySize) {
         this.partitioningBySize = partitioningBySize;
         return this;
@@ -37,6 +51,16 @@ public class PartitioningSettings {
 
     public PartitioningSettings setPartitionSize(long partitionSizeMb) {
         this.partitionSizeMb = partitionSizeMb;
+        return this;
+    }
+
+    public PartitioningSettings setMinPartitionsCount(long partitionsCount) {
+        this.minPartitionsCount = partitionsCount;
+        return this;
+    }
+
+    public PartitioningSettings setMaxPartitionsCount(long partitionsCount) {
+        this.maxPartitionsCount = partitionsCount;
         return this;
     }
 }